refactor: Clean up existing debug functions.
Deleted debug function to upload local HF set file as this can be done by sending a message to local service bus topic, as described in the Developer Zero To Hero Guide.
