+++ Atlas.Common/Utils/Http/RouteConstants.cs
@@ -1,0 +1,7 @@
+﻿namespace Atlas.Common.Utils.Http
+{
+    public static class RouteConstants
+    {
+        public const string DebugRoutePrefix = "debug";
+    }
+}
